+++ lib/utils/transformers/weeklySchedule.js
@@ -1,0 +1,2 @@
+import { toWeeklySchedules } from '../../../src/utils/transformers/weeklySchedules'
+export { toWeeklySchedules }
