+++ src/components/Modal.vue
@@ -17,7 +17,7 @@
 
 <script>
 import { mixin as clickaway } from 'vue-clickaway'
-import { freezeBackground, unfreezeBackground } from '../../lib/utils/modals'
+import { freezeBackground, unfreezeBackground } from '../utils/modals'
 import '../scss/Modal.scss'
 
 export default {
