+++ src/utils/colors.js
@@ -1,0 +1,18 @@
+export const avatarColors = ['#5A8770', '#B2B7BB', '#6FA9AB', '#F5AF29', '#0088B9', '#F18636', '#D93A37', '#A6B12E', '#5C9BBC', '#F5888D', '#9A89B5', '#407887', '#9A89B5', '#5A8770', '#D33F33', '#A2B01F', '#F0B126', '#0087BF', '#F18636', '#0087BF', '#B2B7BB', '#72ACAE', '#9C8AB4', '#5A8770', '#EEB424', '#407887']
+export const wageChartColors = ['#5cbae6', '#1f78b4', '#b6d957', '#33a02c', '#fb9a99', '#e31a1c', '#fdbf6f', '#ff7f00', '#6a3d9a', '#fac364', '#b15928']
+export const extraWageColors = ['#69d2e7', '#a7dbd8', '#e0e4cc', '#f38630', '#fa6900', '#fe4365', '#fc9d9a', '#f9cdad', '#c8c8a9', '#83af9b', '#ecd078', '#d95b43', '#c02942', '#542437', '#53777a', '#556270', '#4ecdc4', '#c7f464', '#ff6b6b', '#c44d58', '#e08e79', '#f1d4af', '#ece5ce', '#c5e0dc', '#e8ddcb', '#cdb380', '#036564', '#490a3d', '#bd1550', '#e97f02', '#f8ca00', '#8a9b0f', '#594f4f', '#547980', '#45ada8', '#9de0ad', '#e5fcc2', '#00a0b0', '#6a4a3c', '#cc333f', '#eb6841', '#edc951', '#e94e77', '#d68189', '#c6a49a', '#c6e5d9', '#f4ead5', '#3fb8af', '#7fc7af', '#dad8a7', '#ff9e9d', '#ff3d7f', '#d5ded9', '#7a6a53', '#99b2b7', '#cbe86b', '#555152', '#99173c', '#005f6b', '#008c9e', '#00b4cc', '#00dffc', '#73626e', '#b38184', '#f0b49e', '#f7e4be', '#ff4e50', '#fc913a', '#ede574', '#e1f5c4', '#99b898', '#fecea8']
+
+export const calendarColors = [
+  { hex: '#a24689', name: 'Default' },
+  { hex: '#0a7d40', name: 'Basil' },
+  { hex: '#35b579', name: 'Sage' },
+  { hex: '#029ce3', name: 'Peacock' },
+  { hex: '#405ea8', name: 'Blueberry' },
+  { hex: '#7986c9', name: 'Lavender' },
+  { hex: '#8b23a8', name: 'Grape' },
+  { hex: '#FF0080', name: 'Flamingo' }
+]
+export const calendarGray = '#BBBBBB'
+export const dayOffEducationColor = '#f5be27'
+export const dayOffSickDayColor = '#d40000'
+export const companyHolidayColor = '#692192'
